[⚙️ chore] margin, padding rem 단위 변경 및 공통 타입 분리 #1
