+++ src/pages/register/index.tsx
@@ -2,27 +2,18 @@ import styled from '@emotion/styled';
 import { useState } from 'react';
 import { useRouter } from 'next/router';
 import Image from 'next/image';
-import media from '@/styles/media';
-import screenSize from '@/constants/screenSize';
 import Header from '@/components/common/Header';
 import Btn from '@/components/common/Btn';
 import ServiceTermCheck from '@/components/register/ServiceTermCheck';
 import PrivacyInput from '@/components/register/PrivacyInput';
 import NicknameInput from '@/components/register/NicknameInput';
 import JobTitleInput from '@/components/register/JobTitleInput';
-import { TGenderOrNull } from '@/types/gender';
-
-export interface IRegisterState {
-  termAgreement: boolean;
-  birthYear: string;
-  gender: TGenderOrNull;
-  nickname: string;
-  jobTitle: string;
-}
+import { IRegisterState } from '@/types/register';
+import { SLayoutWrapper } from '@/components/common/Layout';
 
 export default function Register() {
   const router = useRouter();
-  const [step, setStep] = useState(1);
+  const [step, setStep] = useState<number>(1);
   const [registerData, setRegisterData] = useState<IRegisterState>({
     termAgreement: false,
     birthYear: '',
@@ -85,126 +76,120 @@ export default function Register() {
 
   return (
     <SRegisterWrapper>
-      <Header headerText="회원가입" />
-      <SRegisterBackBtn onClick={goToPreviousStep}>
-        <Image
-          src="/icons/icon-left-arrow.svg"
-          width={24}
-          height={24}
-          alt="뒤로가기 버튼"
-        />
-      </SRegisterBackBtn>
-      <h2 className="a11yHidden">회원가입</h2>
-      <SRegisterStepGuide step={step}>
-        <SCurrentStep step={step}>({step}/4)</SCurrentStep>
-        {step === 1 && (
-          <h3>
-            서비스 이용 약관에
-            <br /> 동의해주세요.
-          </h3>
-        )}
-        {step === 2 && (
-          <h3>
-            회원님의 정보를
-            <br /> 입력해주세요.
-          </h3>
-        )}
-        {step === 3 && <h3>닉네임을 입력해주세요.</h3>}
-        {step === 4 && <h3>해당하는 직군을 선택해주세요.</h3>}
-      </SRegisterStepGuide>
-      <form action="" method="post" name="registerForm">
-        {step === 1 && (
-          <ServiceTermCheck updateRegisterData={updateRegisterData} />
-        )}
-        {step === 2 && (
-          <PrivacyInput
-            gender={registerData.gender}
-            updateRegisterData={updateRegisterData}
+      <main>
+        <Header headerText="회원가입" />
+        <SRegisterBackBtn onClick={goToPreviousStep}>
+          <Image
+            src="/icons/icon-left-arrow.svg"
+            width={24}
+            height={24}
+            alt="뒤로가기 버튼"
           />
-        )}
-        {step === 3 && (
-          <NicknameInput updateRegisterData={updateRegisterData} />
-        )}
-        {step === 4 && (
-          <JobTitleInput
-            updateRegisterData={updateRegisterData}
-            jobTitle={registerData.jobTitle}
-          />
-        )}
-      </form>
-      <SRegisterNextBtnWrapper onClick={goToNextStep}>
-        {step === 1 && (
-          <Btn
-            btns={[
-              {
-                text: '다음',
-                styleType:
-                  step === 1 && !registerData.termAgreement
-                    ? 'disabled'
-                    : 'primary',
-                size: 'large',
-              },
-            ]}
-          />
-        )}
-        {step === 2 && (
-          <Btn
-            btns={[
-              {
-                text: '다음',
-                styleType:
-                  step === 2 && !registerData.birthYear
-                    ? 'disabled'
-                    : 'primary',
-                size: 'large',
-              },
-            ]}
-          />
-        )}
-        {step === 3 && (
-          <Btn
-            btns={[
-              {
-                text: '다음',
-                styleType:
-                  step === 3 && !registerData.nickname ? 'disabled' : 'primary',
-                size: 'large',
-              },
-            ]}
-          />
-        )}
-        {step === 4 && (
-          <Btn
-            btns={[
-              {
-                text: '완료',
-                styleType:
-                  step === 4 && !registerData.jobTitle ? 'disabled' : 'primary',
-                size: 'large',
-              },
-            ]}
-          />
-        )}
-      </SRegisterNextBtnWrapper>
+        </SRegisterBackBtn>
+        <h2 className="a11yHidden">회원가입</h2>
+        <SRegisterStepGuide step={step}>
+          <SCurrentStep step={step}>({step}/4)</SCurrentStep>
+          {step === 1 && (
+            <h3>
+              서비스 이용 약관에
+              <br /> 동의해주세요.
+            </h3>
+          )}
+          {step === 2 && (
+            <h3>
+              회원님의 정보를
+              <br /> 입력해주세요.
+            </h3>
+          )}
+          {step === 3 && <h3>닉네임을 입력해주세요.</h3>}
+          {step === 4 && <h3>해당하는 직군을 선택해주세요.</h3>}
+        </SRegisterStepGuide>
+        <form action="" method="post" name="registerForm">
+          {step === 1 && (
+            <ServiceTermCheck updateRegisterData={updateRegisterData} />
+          )}
+          {step === 2 && (
+            <PrivacyInput
+              gender={registerData.gender}
+              updateRegisterData={updateRegisterData}
+            />
+          )}
+          {step === 3 && (
+            <NicknameInput updateRegisterData={updateRegisterData} />
+          )}
+          {step === 4 && (
+            <JobTitleInput
+              updateRegisterData={updateRegisterData}
+              jobTitle={registerData.jobTitle}
+            />
+          )}
+        </form>
+        <SRegisterNextBtnWrapper onClick={goToNextStep}>
+          {step === 1 && (
+            <Btn
+              btns={[
+                {
+                  text: '다음',
+                  styleType:
+                    step === 1 && !registerData.termAgreement
+                      ? 'disabled'
+                      : 'primary',
+                  size: 'large',
+                },
+              ]}
+            />
+          )}
+          {step === 2 && (
+            <Btn
+              btns={[
+                {
+                  text: '다음',
+                  styleType:
+                    step === 2 && !registerData.birthYear
+                      ? 'disabled'
+                      : 'primary',
+                  size: 'large',
+                },
+              ]}
+            />
+          )}
+          {step === 3 && (
+            <Btn
+              btns={[
+                {
+                  text: '다음',
+                  styleType:
+                    step === 3 && !registerData.nickname
+                      ? 'disabled'
+                      : 'primary',
+                  size: 'large',
+                },
+              ]}
+            />
+          )}
+          {step === 4 && (
+            <Btn
+              btns={[
+                {
+                  text: '완료',
+                  styleType:
+                    step === 4 && !registerData.jobTitle
+                      ? 'disabled'
+                      : 'primary',
+                  size: 'large',
+                },
+              ]}
+            />
+          )}
+        </SRegisterNextBtnWrapper>
+      </main>
     </SRegisterWrapper>
   );
 }
 
-const SRegisterWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: ${screenSize.max}px;
-  height: 100vh;
-  margin: 0 auto;
-  background-color: #fff;
-  box-shadow: 0 0 20px rgba(0, 0, 0, 0.1);
-  @supports (-webkit-touch-callout: none) {
-    height: -webkit-fill-available;
+const SRegisterWrapper = styled(SLayoutWrapper)`
+  & main {
   }
-  ${media.mobileMax} {
-    width: 100vw;
-  }
-  position: relative;
 `;
 
 const SRegisterBackBtn = styled.button`
@@ -223,7 +208,8 @@ const SRegisterStepGuide = styled.div<{ step: number }>`
   h3 {
     line-height: 1.4;
     height: ${({ step }) => (step === 3 || step === 4 ? '28px' : '56px')};
-    margin-bottom: ${({ step }) => (step === 3 || step === 4 ? '8px' : '16px')};
+    margin-bottom: ${({ step }) =>
+      step === 3 || step === 4 ? '.5rem' : '1rem'};
     margin-top: 2rem;
     margin-left: 1.25rem;
     font-size: ${({ theme }) => theme.fontSize.headline2};
