+++ src/types/register.ts
@@ -1,0 +1,9 @@
+import { TGenderOrNull } from './gender';
+
+export interface IRegisterState {
+  termAgreement: boolean;
+  birthYear: string;
+  gender: TGenderOrNull;
+  nickname: string;
+  jobTitle: string;
+}
